Updated to remove warnings in VisualStudio
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/RobWork/trunk@1257 494c09c7-a67d-49d4-9938-151fb1965be8

+++ src/rw/loaders/xml/XMLPropertyLoader.cpp
@@ -68,7 +68,7 @@ namespace {
        case PropertyType::String:
            return ownedPtr(new Property<std::string>(name, description, XMLBasisTypes::readString(child)));
        case PropertyType::Float:
-           return ownedPtr(new Property<float>(name, description, XMLBasisTypes::readDouble(child)));
+           return ownedPtr(new Property<float>(name, description, (float)XMLBasisTypes::readDouble(child)));
        case PropertyType::Double:
            return ownedPtr(new Property<double>(name, description, XMLBasisTypes::readDouble(child)));
        case PropertyType::Int:
